+++ packages/node-core/src/configure/configure.module.ts
@@ -52,7 +52,7 @@ export function yargsToIConfig(yargs: Args, nameMapping: Record<string, string> 
   return Object.entries(yargs).reduce((acc, [key, value]) => {
     if (['_', '$0'].includes(key)) return acc;
 
-    if (key === 'network-registry') {
+    if (isMatchFlag(key, ['network-registry'])) {
       try {
         value = JSON.parse(value as string);
       } catch (e) {
@@ -61,7 +61,7 @@ export function yargsToIConfig(yargs: Args, nameMapping: Record<string, string> 
     }
 
     // Merge network endpoints and possible endpoint configs
-    if (key === 'network-endpoint') {
+    if (isMatchFlag(key, ['network-endpoint'])) {
       const endpointConfig = processEndpointConfig(yargs['network-endpoint-config']);
       if (typeof value === 'string') {
         value = [value];
@@ -76,22 +76,26 @@ export function yargsToIConfig(yargs: Args, nameMapping: Record<string, string> 
         );
       }
     }
-    if (key === 'primary-network-endpoint') {
+    if (isMatchFlag(key, ['primary-network-endpoint'])) {
       const endpointConfig = processEndpointConfig(yargs['primary-network-endpoint-config']);
       value = [value, endpointConfig[0] ?? {}];
     }
-    if (['network-endpoint-config', 'primary-network-endpoint-config'].includes(key)) return acc;
+    if (isMatchFlag(key, ['network-endpoint-config', 'primary-network-endpoint-config'])) return acc;
 
-    if (key === 'disable-historical' && value) {
+    if (isMatchFlag(key, ['disable-historical']) && value) {
       acc.historical = false;
     }
-    if (key === 'historical' && value === 'false') {
+    if (isMatchFlag(key, ['historical']) && value === 'false') {
       value = false;
     }
 
     acc[nameMapping[key] ?? camelCase(key)] = value;
     return acc;
   }, {} as any);
+}
+
+function isMatchFlag(key: string, targets: string[]): boolean {
+  return targets.some((target) => key === target || key === camelCase(target));
 }
 
 function warnDeprecations(argv: Args) {