+++ packages/node-core/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- When the `primary-network-endpoint` flag exists, it will report an "Invalid endpoint" error.
+
 ## [15.0.3] - 2024-11-26
 ### Fixed
 - Workers crashing because of lazy monitor write (#2607)